Replace movePlayer with a Command. Stop overwriting existing commands in queue when closing UI.
