+++ flows/commands/commands.ts
@@ -5,11 +5,13 @@ import { CommandDef, Command, Soul } from '../../types';
 import { blastCmd } from './blast-command';
 import { takeCmd } from './take-command';
 import { bonkCmd } from './bonk-command';
+import { moveCmd } from './move-command';
 
 export var cmdDefsById: Record<string, CommandDef> = {
   blast: { id: 'blast', name: 'Blast', cmdFn: blastCmd },
   take: { id: 'take', name: 'Take', cmdFn: takeCmd },
-  bonk: { id: 'bonk', name: 'Bonk', cmdFn: bonkCmd }
+  bonk: { id: 'bonk', name: 'Bonk', cmdFn: bonkCmd },
+  move: { id: 'move', name: 'Move', cmdFn: moveCmd }
 };
 
 export function instantiateCmdFromDef(
